Add two new metrics

+++ speedy-murmurs-simulator/src/main/java/treeembedding/credit/CreditNetwork.java
@@ -548,7 +548,7 @@ public class CreditNetwork extends AbstractCreditNetworkBase {
     for (int treeIndex = 0; treeIndex < mins.length; treeIndex++) {
       NextHopPlusMetrics n = ra.getRoute(src, dest, treeIndex, g, nodes, exclude);
       paths[treeIndex] = n.getPath();
-      addPerEpochValue(Metrics.BLOCKED_LINKS_PER_EPOCH, (double) n.getBlockedLinks(), calculateEpoch(cur));
+      addPerEpochValue(Metrics.BLOCKED_LINKS_PER_EPOCH, n.getBlockedLinks(), calculateEpoch(cur));
 
       if (paths[treeIndex][paths[treeIndex].length - 1] == dest) {
         int currentNodeIndex = src;
